### 2021-04-03 #53 @NishantPrabhu
[Re] On the Relationship between Self-Attention and Convolutional Layers
**Original article:** *On the Relationship between Self-Attention and Convolutional Layers* by Jean-Baptiste Cordonnier, Andreas Loukas & Martin Jaggi, in proceedings of ICML 2020.

**PDF URL:**  [Reproducibility report](https://github.com/NishantPrabhu/Self-Attention-and-Convolutions/blob/main/article.pdf)
**Metadata URL:** [Metadata](https://github.com/NishantPrabhu/Self-Attention-and-Convolutions/blob/main/metadata.tex)
**Code URL:** [Code](https://github.com/NishantPrabhu/Self-Attention-and-Convolutions)

**Scientific domain:** Deep Learning
**Suggested editor:** Nicolas P. Rougier (@rougier)

This is a reproduction of the paper mentioned above, written with Mukund Varma (@mukundvarmat). The paper discusses the use of transformers and self-attention for vision tasks. We suggest Nicolas P. Rougier as the editor.

With warm regards,
Nishant Prabhu
