### 2021-04-08 #53 @NishantPrabhu
@rougier

Greetings!

1. We intend our submission to be evaluated as a regular ReScience submission. The current submission has already been redesigned according to the template prescribed on the repository. If any other form of adaptation/modification is desired, please let us know and we will get it done as soon as possible.

2. We had gone through and responded to the reviews given during MLRC 2020, and our submission had been revised to concur with the suggestions in the best possible manner. Our submission here contains the revised content.

Please let us know if there's anything else we should attend to. Thanks!
