### 2021-04-29 #53 @don-tpanic
Hi @NishantPrabhu

I ran into this error when running `python3 main.py --config 'configs/learned_2d.yaml' --task 'train'` inside `main/`

```
Traceback (most recent call last):
  File "/home/ken/projects/Self-Attention-and-Convolutions/main/main.py", line 414, in <module>
    trainer = Trainer(args)
  File "/home/ken/projects/Self-Attention-and-Convolutions/main/main.py", line 28, in __init__
    self.encoder = networks.Encoder(self.config['encoder']).to(self.device)
AttributeError: module 'models.networks' has no attribute 'Encoder'
```

I suspect this is something minor like the attribute was renamed? Could you check this for me and update the code?

Thanks!
Ken
