### 2021-04-29 6252b31
Fixed misnamed Encoder class
