### 2021-04-29 #53 @NishantPrabhu
Hi @don-tpanic

Thanks for pointing out this issue. As you rightly diagnosed, a class name had been modified. I've fixed it now and it's working fine at my end. Let me know if you face any other issues.

Thanks and with warm regards,
Nishant.
