### 2021-04-29 #53 @NishantPrabhu
Hi again @don-tpanic

If you're planning to train the model for a while, you might want to modify the batch size in the config files to suit your accelerator's capacity. I've set it to 1 for testing the changes I recently made.

Thanks,
Nishant.
