### 2021-04-29 #53 @don-tpanic
Thanks @NishantPrabhu for your prompt response!

Could you look into this error for me?

When I ran ```python3 main.py --config 'configs/resnet.yaml' --task 'train'```

```
Traceback (most recent call last):
  File "/home/ken/projects/Self-Attention-and-Convolutions/main/main.py", line 609, in <module>
    trainer = Trainer(args)
  File "/home/ken/projects/Self-Attention-and-Convolutions/main/main.py", line 38, in __init__
    self.encoder = networks.Encoder(self.config['encoder']).to(self.device)
  File "/home/ken/projects/Self-Attention-and-Convolutions/main/models/networks.py", line 109, in __init__
    self.num_blocks = config['num_encoder_blocks']
KeyError: 'num_encoder_blocks'
```

Thanks,
Ken
