### 2021-04-29 d40c15e
main script some changes
