### 2021-04-29 12bcbce
Readme updated with CLI instructions
