+++ README.md
@@ -46,3 +46,25 @@ To visualize this trained model's attention as an average over all pixels of the
 ```
 python3 main.py --config 'configs/learned_2d.yaml' --task 'viz' --load 'output/dir/with/trained/model'
 ```
+
+### Task options
+The `--task` argument for different functionalities has been listed below.
+
+**For paper's experiments (`main`)**
+  - `train`: Train any of the models except a ResNet model.
+  - `resnet_train`: Train a ResNet model.
+  - `viz_1`: Visualize attention patterns for a single query pixel in an image. The chosen image is currently the first one generated by the validation dataloader.
+  - `viz`: Visualize average attention patterns in an image. The chosen image is currently the first one generated by the validation dataloader.
+  - `gauss_viz`: Visualize attention centers of quadratic embedding based model.
+  - `time`: Compute the average throughput and inference time of the model.
+
+**For SAN experiments (`SAN`)**
+  - `train`: Train any of the models.
+  - `viz`: Visualize average attention patterns for an image. To use this, please create a directory `imgs/` inside `SAN/` and add one image to it.
+  - `time`: Compute the average throughput and inference time of the model.
+
+**For Vision Transformer experiments (`ViT`)**
+  - `train`: Train any of the models except a ResNet model.
+  - `viz_1`: Visualize attention patterns for a single query pixel in an image. To use this, please create a directory `imgs/` inside `ViT/` and add one image to it.
+  - `viz`: Visualize average attention patterns in an image. To use this, please create a directory `imgs/` inside `ViT/` and add one image to it.
+  - `time`: Compute the average throughput and inference time of the model.
