### 2021-04-29 #53 @NishantPrabhu
Hi @don-tpanic

The CLI command to train a ResNet model is slightly different. I understand the confusion since we haven't mentioned this detail in the README, which I have added now. Please check the **Task options** section in the README to know more.

In particular, to train a ResNet model, the correct command would be:

```
python3 main.py --config 'configs/resnet.yaml' --task 'resnet_train'
```

Note that `--task` is given `resnet_train` and not `train`. It is so because our trainer definitions for the two are quite different.

Let me know in case of other issues. Thanks!

Nishant.
