### 2021-04-30 #53 @don-tpanic
Hi @NishantPrabhu,

Thanks for updating the README. I was able to run training with resnet perfectly!

Here is another error I came across when training quadratic embedding with isotropic gaussian.
`python3 main.py --config 'configs/gaussian.yaml' --task 'train'`

The error was thrown at the end of validation:
```
[TRAIN] Epoch [  1/300]
Progress: [===================>] 100.00% [Loss] 1.8334 [Accuracy] 0.3218
[VALID] Epoch [  1/300]
Progress: [===================>] 100.00% [Loss] 1.5961 [Accuracy] 0.4208 Traceback (most recent call last):
  File "/home/ken/projects/Self-Attention-and-Convolutions/main/main.py", line 610, in <module>
    trainer.train()
  File "/home/ken/projects/Self-Attention-and-Convolutions/main/main.py", line 438, in train
    self.visualize_attention(epoch)
  File "/home/ken/projects/Self-Attention-and-Convolutions/main/main.py", line 197, in visualize_attention
    ax.imshow(img_normal.permute(1, 2, 0).cpu().numpy())
RuntimeError: number of dims don't match in permute
```
I believe the error was originated in `visualize_attention`. One related question if you could clarify it for me: it seems attention will always be visualized no matter if the task is `train` or `viz`? I wonder if this is intended behaviour or could be sligtly improved?

Thanks,
Ken
