+++ README.md
@@ -53,8 +53,8 @@ The `--task` argument for different functionalities has been listed below.
 **For paper's experiments (`main`)**
   - `train`: Train any of the models except a ResNet model.
   - `resnet_train`: Train a ResNet model.
-  - `viz_1`: Visualize attention patterns for a single query pixel in an image. The chosen image is currently the first one generated by the validation dataloader.
-  - `viz`: Visualize average attention patterns in an image. The chosen image is currently the first one generated by the validation dataloader.
+  - `viz_1`: Visualize attention patterns for a single query pixel in an image. To use this, please create a directory `imgs/` inside `ViT/` and add one image to it.
+  - `viz`: Visualize average attention patterns in an image. To use this, please create a directory `imgs/` inside `ViT/` and add one image to it.
   - `gauss_viz`: Visualize attention centers of quadratic embedding based model.
   - `time`: Compute the average throughput and inference time of the model.
 
