### 2021-04-30 #53 @NishantPrabhu
Hi @don-tpanic

First of all, I'm really sorry about all these roadblocks you're facing. Looks like we didn't handle some of our code correctly after we modified it post our RC2020 reviews.

Anyway, I have updated the code to improve consistency across experiments. When you use set `--task` as `viz`, the code will now look for `imgs/` in the working directory and use one image from it for visualization. A folder with that name is already present in the experiment directories, with some handpicked images in it. Please note that for SAN and ViT, loading a pretrained model using `--load` is necessary for visualization to work. For `main`, it will work but will generate useless results.

Also, I've disabled visualization during training. We had done that to observe the evolution of attention patterns with training, which should not be necessary now. If needed in the future, just uncomment line 433 in `main/main.py`. This part of the code for all three experiments is now working fine on my end.

Thanks and with warm regards,
Nishant.
