### 2021-05-26 #53 @nicksexton
Thanks for your patience, we have had a chance to confer on this. We have worked through running the models in table 1 and generally, everything checks out.

1. For table 1, we have an issue with the final two models (Learned embeddings w/ and w/o content) - we find this exceeds the memory on our GPUs (RTX 2080). We reduced the batch size accordingly (from 100 to 32) however with this setting we cannot get the model to train. We are unsure of the best way to proceed on this, @oliviaguest this seems to be an interesting reproducibility question, wonder if you could advise. It would be helpful to have more detail on the computational setup used? We're unsure whether we actually have any GPUs in lab big enough to run this model with the specified batch size without falling back on e.g. AWS.
2. Regarding (e.g.) figure 1, is there a script which automates parameterising, training and warehousing data from all the models in order produce the figure? Currently this seems to require individual runs of the model with specific parameter settings (the various models do not 100% match the naming of the configs used to run the experiments. what are w/ and w/o content in the experiment config? is it the `learned_2d.ymal` - `use_attention_data`?), we have been able to make educated guesses about these, but for reproducibility this needs to be explicitly stated (ideally in a script, or in the readme).
3. The part of the data processing pipeline that encompasses retrieving stored simulation results and generating figures relies on an undocumented jupyter notebook (`./temp/log_extract.ipynb`). Not that there is anything wrong with notebooks per se but we are specifically concerned about the need to manually configure model runs and name datasets (2, above) where specific filenames are hard-coded in the plotting notebook. As a minimum, we think there is a need for documented code that automates this part of the process.
4. Currently, as far as table 1, all our reproductions check out and we are reasonably confident that any given result in the paper will generate the results stated. However, although sufficient information is provided that we can run any given model, lack of high-level automated scripts makes reproducing a given table or figure quite a manual process that is a) potentially subject to user error, and b) consuming significant amounts of time on our end. We wonder if there can be a discussion of what is reasonable here @oliviaguest. Please could you advise if we should be checking all models reported in the paper or if spot checks are sufficient? If we are aiming for reproducing all results in the paper, can we ask ideally for simulation scripts so that running each figure or table is a one-liner.

Thanks! Nick
