### 2021-05-26 #53 @oliviaguest
Thanks for all this, you @nicksexton and @don-tpanic! ☺️

Two main points/takeaways from me:

*a*. @NishantPrabhu, you now probably have a lot of small and large changes to make to improve this replication. Feel free to deal with these while everybody else discusses *b* (below), if you want.

*b*. I am not sure about the questions above (in points 1 and 4). @khinsen @benoit-girard @rougier can you chime in too, please?  I would like the author(s; @NishantPrabhu) to also give their perspective on this, when possible.
