### 2021-05-26 #53 @MukundVarmaT
Hi @nicksexton
Thank you for reviewing our reproducibility report.

1. We would like to direct you to section 3.2 in the submitted reproducibility report which mentions the required computational resources to replicate all the results. To summarize, the learned embedding-based experiments were run on 16GB NVIDIA V100 GPUs rented from AWS, and the others were run on an 8GB NVIDIA RTX 2060 owned by us.
2. We apologize for our oversight in not including individual config files for each training experiment and we shall include them as soon as possible for convenience. This should also improve consistency in naming schemes between the paper and the code.
3. The provided code does include methods/functions to visualize the attention plots. With regards to fig. 1 in the reproducibility report (performance vs epochs across various models), we have used a notebook to extract the required information from the corresponding log files. We would also like to point out that wandb (which is used for logging our experiments) does allow for comparison across various training runs which should produce the same figure. Therefore we have not included a separate script for the same.

Thoughts about 1. and 4.
Thank you for validating and confirming our results in Table 1.  As mentioned in 2., we will include separate configuration files for each run which should help automate reproducing all the results. Indeed some of these experiments are quite computationally heavy making it quite difficult to validate all our results. If it helps, we could provide our training logs from wandb as a check for our claimed results (which could be an alternative from having to run everything from your end).

PS: As mentioned in our report, we go beyond reproducibility and propose a new variant of the attention operation (to possibly overcome some of the problems observed during reproducibility).  Since the discussion (so far) didn't mention it, we wanted to highlight it to the reviewers and would love to hear their thoughts regarding the same.

Thanks and Regards,
Mukund and Nishant
