+++ main/models/attention.py
@@ -192,7 +192,7 @@ class Learned2dRelativeSelfAttention(nn.Module):
             self.value = nn.Linear(self.model_dim, self.model_dim*self.heads, bias=False)
             self.out = nn.Linear(self.model_dim*self.heads, self.model_dim)
         elif self.heads == 9:
-            self.value = nn.Linear(self.model_dim*self.heads, self.model_dim, bias=False)
+            self.value = nn.Linear(self.model_dim*self.heads, self.model_dim)
             
         self.dropout = nn.Dropout(config.get('attention_dropout_prob', 0.1))
         self.layer_norm = nn.LayerNorm(self.model_dim)
@@ -317,7 +317,7 @@ class GaussianAttention(nn.Module):
         self.attention_spreads = nn.Parameter(attention_spreads)
 
         # Other layers
-        self.out = nn.Linear(self.heads*self.model_dim, self.model_dim, bias=True)
+        self.value = nn.Linear(self.heads*self.model_dim, self.model_dim, bias=True)
         self.layer_norm = nn.LayerNorm(self.model_dim)
 
         # If not using gaussian blur trick, define quadratic positional encoding
@@ -405,7 +405,7 @@ class GaussianAttention(nn.Module):
         else:
             context = self.blurred_attention(hidden_state)                                          # (bs, h, w, heads*model_dim)
 
-        output = self.out(context)                                                                # (bs, h, w, model_dim)
+        output = self.value(context)                                                                # (bs, h, w, model_dim)
 
         # Residual connection
         output = output + hidden_state
