### 2021-05-27 #53 @NishantPrabhu
Greetings!

I have added experiment specific configuration files to our repo and updated the README with more information about them. Please note that I have only included configurations of experiments we performed, even though more combinations are possible. Kindly let me know if anything else needs to be done.

Thanks and with warm regards,
Nishant.
