### 2021-05-27 909df51
Vanilla SAN configs added, README updated
