+++ SAN/main.py
@@ -75,7 +75,7 @@ class Trainer:
             self.logger.record("No saved state found. Starting fresh", mode="info")
 
         # Load best model if any argument is provided
-        if 'load' in args.keys():
+        if args["load"] is not None:
             if os.path.exists(os.path.join(args['load'], 'best_model.ckpt')):
                 self.load_model(args['load'])
                 self.logger.print(f"Successfully loaded model at {args['load']}", mode='info')
