### 2021-06-07 #53 @don-tpanic
Hi @NishantPrabhu,

I have managed to reproduce more results from the paper and things are looking really good so far. Here are some minor suggestions which I think could improve the ease of reproducibility for others in the future.

1. For the moment, in order to visualise attention over images, one has to trace back into the source code for where to supply the image file name. It would be nice if this bit is less hard-coded.

2. Again, minor suggestion but I think the command provided for visualising attention could be improved. The current command could potentially cause 2 problems when no `--output` is provided by the user. i) as it defaults to a directory named based on the current timestamp, it might lead to confusion in that model checkpoints and associated results are located in different directories (took me a while to find the nice attention plots); ii) when using this command for visualising gaussians in Figure 2 (with `--task 'gauss_viz'`), leaving `--output` to default will actually be a problem because the code looks into this new directory based on timestamp that does not have any checkpoints.

Basically, I suggest either making this more explicit that both `--output` and `--load` need provided and consistent or it would be better when providing `--load` is sufficient.

3. In our previous feedback, Nick mentioned that currently the code for plotting Figure 1 is a jupyter notebook. As he suggested, perhaps it would be better to have a standard script for this part like everything else. There doesn't seem to be enough information for using this notebook either in the README.  Where this notebook is located can also be a bit hard to find as the directory is named `temp`.

Best,
Ken
