### 2021-06-07 f5ae829
Visualization functionality improved
