+++ ViT/main.py
@@ -81,6 +81,7 @@ class Trainer:
         if 'load' in args.keys():
             if os.path.exists(os.path.join(args['load'], 'best_model.ckpt')):
                 self.load_model(args['load'])
+                aelf.output_dir = args['load']
                 self.logger.print(f"Succesfully loaded model from {args['load']}")
             else:
                 self.logger.print(f"No saved model found at {args['load']}; please check your input!")
@@ -223,7 +224,7 @@ class Trainer:
         
         # Load image
         transform = T.Compose([T.Resize(32), T.ToTensor(), T.Normalize([0.4914, 0.4822, 0.4465], [0.2023, 0.1994, 0.2010])])
-        img = Image.open('imgs/{}'.format(os.listdir("imgs")[0]))
+        img = Image.open(self.args["image"])
         img = transform(img).unsqueeze(0)
 
         for k, idx in enumerate([16, 17]):
@@ -273,8 +274,7 @@ class Trainer:
                     count += 1
 
             plt.tight_layout(pad=0.5)
-            # plt.show()
-            plt.savefig('/home/nishant/Desktop/vit_2.pdf', pad_inches=0.05)
+            plt.savefig(os.path.join(self.output_dir, "viz_image.png"), pad_inches=0.05)
 
 
     def train(self):
@@ -336,6 +336,7 @@ if __name__ == '__main__':
     ap.add_argument('-c', '--config', required=True, help='Path to configuration file')
     ap.add_argument('-o', '--output', default=dt.now().strftime('%Y-%m-%d_%H-%M'), type=str, help='Path to output file')
     ap.add_argument('-l', '--load', type=str, help='Directory from which best model will be loaded')
+    ap.add_argument('-i', '--image', default="imgs/cat_s_000019.png", help='Path to image on which attention visualization will be performed')
     ap.add_argument('-t', '--task', type=str, required=True, help='The task to perform, choose from (train, viz, viz_1, time)')
     args = vars(ap.parse_args())
 
@@ -347,11 +348,18 @@ if __name__ == '__main__':
         trainer.train()
 
     elif args['task'] == 'viz':
+        if args['load'] is None:
+            raise NotImplementedError("Please load a model using --load for visualization tasks")
         trainer.visualize_attention(1, False)
 
     elif args['task'] == 'viz_1':
+        if args['load'] is None:
+            raise NotImplementedError("Please load a model using --load for visualization tasks")
         trainer.visualize_attention(1, True)
 
     elif args['task'] == 'time':
         trainer.throughput()
         trainer.inference_time()
+
+    else:
+        raise ValueError(f"Unrecognized task {args['task']}, choose from ('train', 'viz', 'viz_1', 'time')")
