### 2021-06-07 #53 @NishantPrabhu
Hi @don-tpanic,

Thanks for the suggestions!

1. I have added another parameter `--image` to the list of CLI arguments where you can specify the path to an image on which visualization operations will be performed. I've added information about this option to the README as well.
2. I've modified the code so that whenever a directory is specified using `--load`, the output directory will be overwritten with this argument. Thus, all visualization outputs will be stored in the same directory containing the trained model. I hope this is what was expected.
3. I'll need a little more time to develop a general script for this. I will add it to the repo as soon as I can. Thanks for your patience!

I have tested the code after modifications (1) and (2), and it's working fine on my end. If you face any issues running the code on your end, please let me know.

Please also note the following changes:
1. Not specifying `--load` for visualization tasks in any of the experiments will now throw a `NotImplementedError`. Earlier, it would generate outputs with a randomly initialized model. I found this more sensible as it will notify the user of this requirement instead of generating wrong results.

Thanks and with warm regards,
Nishant Prabhu
