### 2021-06-08 87d9236
Added performance plotting scripts
