### 2021-06-08 #53 @NishantPrabhu
Hi everyone,

I've added a script `plot_performance.py` to plot the train/validation metrics of desired experiments. I've added instructions to use the script in the README. Please let me know if it fulfills the required functionality, or if any issues are faced.

Thanks and with warm regards,
Nishant Prabhu
