### 2021-06-16 #53 @don-tpanic
Hi @NishantPrabhu,

Thanks for all the prompt updates. Here is another issue when I ran `python3 main.py --config 'configs/vit_hier.yaml' --task 'train'`
```Traceback (most recent call last):
  File "/home/ken/projects/Self-Attention-and-Convolutions/ViT/main.py", line 346, in <module>
    trainer = Trainer(args)
  File "/home/ken/projects/Self-Attention-and-Convolutions/ViT/main.py", line 82, in __init__
    if os.path.exists(os.path.join(args['load'], 'best_model.ckpt')):
  File "/home/ken/.conda/envs/rescience/lib/python3.9/posixpath.py", line 76, in join
    a = os.fspath(a)
TypeError: expected str, bytes or os.PathLike object, not NoneType
```

Can you look into it and make sure `SAN/` doesn't have the same issue?

Thanks,
Ken
