### 2021-06-16 3f71628
Fixed checkpoint retrieval bug
