### 2021-06-16 #53 @NishantPrabhu
Greetings!

Sorry about this bug; has been fixed now. Tested okay for `SAN/` and `ViT/` on my end.

Thanks,
Nishant
