### 2021-06-23 #53 @khinsen
@NishantPrabhu Jumping in to ask for your opinion on a [proposal](https://github.com/ReScience/ReScience/issues/104#issuecomment-866317930) for  improving the ReScience C reviewing workflow for future HPC-style submissions. Unrelated to this review, I am not asking you to adopt this scheme retroactively! From your point of view, would such a process be a reasonable burden on authors of future submissions?
