### 2021-07-05 #53 @oliviaguest
Sure, @MukundVarmaT you can make sure your ```metadata.yaml``` file is fully filled in with what is required of you. I need it for the publication script: https://github.com/rescience/articles#15-steps-to-publishing-a-rescience-article. 😊
