+++ metadata.tex
@@ -2,7 +2,8 @@
 
 \def \codeURL{https://github.com/NishantPrabhu/Self-Attention-and-Convolutions}
 \def \codeDOI{}
-\def \dataURL{https://www.cs.toronto.edu/~kriz/cifar.html}
+\def \codeSWH{swh:1:dir:6ab40b1686ee05bc3f9413ced6b1a84c6b203814}
+\def \dataURL{}
 \def \dataDOI{}
 \def \editorNAME{}
 \def \editorORCID{}
@@ -13,27 +14,27 @@
 \def \dateRECEIVED{03 April 2021}
 \def \dateACCEPTED{}
 \def \datePUBLISHED{}
-\def \articleTITLE{[Re] On the relationship between self-attention and convolutional layers}
+\def \articleTITLE{[Re] On the Relationship between Self-Attention and Convolutional Layers}
 \def \articleTYPE{Editorial}
 \def \articleDOMAIN{}
 \def \articleBIBLIOGRAPHY{bibliography.bib}
 \def \articleYEAR{2021}
-\def \reviewURL{}
+\def \reviewURL{https://github.com/ReScience/submissions/issues/53}
 \def \articleABSTRACT{In this report, we perform a detailed study on the paper "On the Relationship between Self-Attention and Convolutional Layers", which provides theoretical and experimental evidence that self attention layers can behave like convolutional layers. The proposed method does not obtain state-of-the-art performance but rather answers an interesting question - \textit{do self-attention layers process images in a similar manner to convolutional layers?} This has inspired many recent works which propose fully-attentional models for image recognition. We focus on experimentally validating the claims of the original paper and our inferences from the results led us to propose a new variant of the attention operation - {\em Hierarchical Attention}. The proposed method shows significantly improved performance with fewer parameters, hence validating our hypothesis. To facilitate further study, all the code used in our experiments are publicly available here\footnote{\url{https://github.com/NishantPrabhu/Self-Attention-and-Convolutions}}.}
-\def \replicationCITE{}
-\def \replicationBIB{}
-\def \replicationURL{}
+\def \replicationCITE{Jean-Baptiste Cordonnier and Andreas Loukas and Martin Jaggi. On the Relationship between Self-Attention and Convolutional Layers. International Conference on Learning Representations.}
+\def \replicationBIB{Cordonnier2020On}
+\def \replicationURL{https://arxiv.org/pdf/1911.03584.pdf}
 \def \replicationDOI{}
 \def \contactNAME{Nishant Prabhu}
 \def \contactEMAIL{me17b084@smail.iitm.ac.in}
-\def \articleKEYWORDS{rescience c, rescience x}
+\def \articleKEYWORDS{rescience c, rescience x, python, pytorch, self-attention}
 \def \journalNAME{ReScience C}
 \def \journalVOLUME{4}
 \def \journalISSUE{1}
 \def \articleNUMBER{}
 \def \articleDOI{}
 \def \authorsFULL{Mukund Varma T and Nishant Prabhu}
-\def \authorsABBRV{M. Varma and N. Prabhu}
+\def \authorsABBRV{M. Varma T and N. Prabhu}
 \def \authorsSHORT{Mukund and Nishant}
 \title{\articleTITLE}
 \date{}
