### 2021-07-05 #53 @MukundVarmaT
@oliviaguest Please find updated [metadata.tex](https://github.com/NishantPrabhu/Self-Attention-and-Convolutions/blob/main/metadata.tex), [article.pdf](https://github.com/NishantPrabhu/Self-Attention-and-Convolutions/blob/main/article.pdf), [metadata.yaml](https://github.com/NishantPrabhu/Self-Attention-and-Convolutions/blob/main/metadata.yaml)

I believe some of the information has to be filled in by you which will also reflect changes onto article.pdf.
