+++ metadata.tex
@@ -20,7 +20,7 @@
 \def \articleBIBLIOGRAPHY{bibliography.bib}
 \def \articleYEAR{2021}
 \def \reviewURL{https://github.com/ReScience/submissions/issues/53}
-\def \articleABSTRACT{In this report, we perform a detailed study on the paper "On the Relationship between Self-Attention and Convolutional Layers", which provides theoretical and experimental evidence that self attention layers can behave like convolutional layers. The proposed method does not obtain state-of-the-art performance but rather answers an interesting question - \textit{do self-attention layers process images in a similar manner to convolutional layers?} This has inspired many recent works which propose fully-attentional models for image recognition. We focus on experimentally validating the claims of the original paper and our inferences from the results led us to propose a new variant of the attention operation - {\em Hierarchical Attention}. The proposed method shows significantly improved performance with fewer parameters, hence validating our hypothesis. To facilitate further study, all the code used in our experiments are publicly available here\footnote{\url{https://github.com/NishantPrabhu/Self-Attention-and-Convolutions}}.}
+\def \articleABSTRACT{In this report, we perform a detailed study on the paper 'On the Relationship between Self-Attention and Convolutional Layers', which provides theoretical and experimental evidence that self attention layers can behave like convolutional layers. The proposed method does not obtain state-of-the-art performance but rather answers an interesting question - do self-attention layers process images in a similar manner to convolutional layers? This has inspired many recent works which propose fully-attentional models for image recognition. We focus on experimentally validating the claims of the original paper and our inferences from the results led us to propose a new variant of the attention operation - Hierarchical Attention. The proposed method shows significantly improved performance with fewer parameters, hence validating our hypothesis. To facilitate further study, all the code used in our experiments are publicly available here - https://github.com/NishantPrabhu/Self-Attention-and-Convolutions.}
 \def \replicationCITE{Jean-Baptiste Cordonnier and Andreas Loukas and Martin Jaggi. On the Relationship between Self-Attention and Convolutional Layers. International Conference on Learning Representations.}
 \def \replicationBIB{}
 \def \replicationURL{https://arxiv.org/pdf/1911.03584.pdf}
