### 2021-07-13 #53 @MukundVarmaT
@oliviaguest It's possible that the figures are very close to the margins (or going into it). Since we wanted to show a side-by-side comparison while still making it sufficiently legible, there was no other option. I am not able to figure out an alternative but we are open to suggestions.
