### 2021-07-13 #53 @oliviaguest
I think you could, *if you wanted* make the text bigger in your figures because regardless of the tiny amount they impinge on the margins, it's small.
