### 2021-07-24 #53 @NishantPrabhu
Hi @oliviaguest

Sorry for the delay. I have updated the PDF such that the images are now (seemingly) inside the margins. The link to the original post now points to the revised PDF. Please have a look and let me know if it works.

Thanks,
Nishant.
