### 2021-08-18 #53 @MukundVarmaT
@oliviaguest
Thank you for the update. I have updated [article.pdf](https://github.com/NishantPrabhu/Self-Attention-and-Convolutions/blob/main/article.pdf) on the repository. I hope this will be alright?
