### 2021-08-27 #53 @oliviaguest
@otizonaizit, are you the right person to ask? Does anybody know how to fix this?

```
(base) ip-145-116-130-74:articles olivia$ ./publish.py --sandbox --metadata metadata.yaml --pdf article.pdf
Uploading content to Zenodo... Traceback (most recent call last):
  File "./publish.py", line 204, in <module>
    upload_content(server, token, article_id, article_file)
  File "./publish.py", line 25, in upload_content
    raise IOError("%s: " % response.status_code + response.json()["message"])
OSError: 404: PID does not exist.
```
