### 2021-09-03 #53 @rougier
I'm a the bus factor! Well, actually you can dot it but I strongly encourage you to test it locally. I already broke it several times and I think Konrad broke it too at some point. Problem is jekyll / jekyll-scholar that only plays nicely with ruby 2.6 (via rbenv).
