### 2021-09-08 #53 @MukundVarmaT
@oliviaguest Hi
Just wanted to confirm if all the formalities for publication are completed, and if there is anything else left to be done?
Thanks
